Integrate with Topic Statistics options Fixes after rebasing with master
