+++ rclcpp/test/test_publisher_subscription_count_api.cpp
@@ -51,6 +51,10 @@ public:
   {
     rclcpp::init(0, nullptr);
   }
+  static void TearDownTestCase()
+  {
+    rclcpp::shutdown();
+  }
 
 protected:
   void SetUp() {}
