+++ rclcpp/test/test_subscription_publisher_count_api.cpp
@@ -44,6 +44,10 @@ public:
   {
     rclcpp::init(0, nullptr);
   }
+  static void TearDownTestCase()
+  {
+    rclcpp::shutdown();
+  }
 
 protected:
   void SetUp() {}
